requirements.txt
###### Requirements without Version Specifiers ######`
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

iso8601
